reduce to gemv unification only

+++ crates/backend-uzu/src/backends/cpu/kernel/ssm/ssd_prefill.rs
@@ -79,7 +79,7 @@ pub fn ssd_prefill_sequential<T: ArrayElement + Float>(
     let channels = channels as usize;
     let head_dim = head_dim as usize;
 
-    let safe_group = group_size.max(1) as usize;
+    let safe_group = group_size.max(1);
 
     unsafe {
         for h in 0..channels {
